[harbour-fire] Fixed compilation error on some platforms
src/FireItem.cpp: In member function 'void FireItem::Private::randomizeThreshold()':
src/FireItem.cpp:213:71: error: no matching function for call to 'qMax(double, const qreal&)'
     iThreshold = qMin(qMax(iThreshold + adjust * (random() - 0.5), min),
                                                                       ^

+++ src/FireItem.cpp
@@ -210,7 +210,7 @@ FireItem::Private::randomizeThreshold()
     const qreal min = minThreshold();
     const qreal adjust = (MaxThreshold - min) / 2;
 
-    iThreshold = qMin(qMax(iThreshold + adjust * (random() - 0.5), min),
+    iThreshold = qMin(qMax(qreal(iThreshold + adjust * (random() - 0.5)), min),
         MaxThreshold);
     HDEBUG(iThreshold);
 }
